scripts/requirements-vibecut.txt
# VibeCut analysis/render Python dependencies
moviepy>=1.0.3
numpy>=1.24.0
opencv-python>=4.9.0
pillow>=10.0.0
pydub>=0.25.1
faster-whisper>=1.1.0
